chore(directive): Rename `ACF` to `Acf`
